requirements.txt
numpy
numpy-stl
open3d